Update DESeq2.Rmd
hi phyloseq dev team :)

I'm seeing a lot of phyloseq users posting questions about complex study designs tagged with DESeq2 on the Bioc support site. There appear to be an ever-increasing number of complex microbiome studies.

It's tedious to respond to them one-by-one so I'm trying to add this note upstream that the open source developers shouldn't be contacted for this kind of statistical support.